requirements.txt
asgiref==3.7.2
common==0.1.2
Django==4.2.7
django-cors-headers==4.3.1
django-subdomains==2.1.0
django-tenants==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysql-client==0.0.1
mysql-connector-python==8.2.0
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
setuptools==69.0.2
sqlparse==0.4.4
